+++ doc/releaseNotes.md
@@ -141,3 +141,7 @@ This file describes the main feature changes for each InfoLogger released versio
 
 ## v2.6.0 - 24/01/2024
 - Updated SWIG version requirement to allow generation of Python3 bindings.
+
+## next
+- Added functions to keep a history of latest log messages: see historyReset() and historyGetSummary().
+- Added function to register error codes descriptions to format message summaries in history.